requirements.txt
httpx==0.26.0
uvicorn==0.27.1
python-dotenv==1.0.1
pytest-cov==6.1.0
fastapi==0.109.2
pytest-asyncio==0.26.0
beautifulsoup4==4.13.0
slowapi==0.1.8